Initial changes for my image
This change adds i386 multi-arch support which is useful for
running 32bit applications.

There is also a build_daviey.sh, which is used to build this
specific fork (based on the offical Kali image, but not running
full debootstrap).

This change also adds an upgrade step at the end, to make sure
the image is latest.

Signed-off-by: Dave Walker (Daviey) <email@daviey.com>